refactor: 확장함수 통합
